se agrega botones pages
