+++ src/AboutApi.js
@@ -1,0 +1,13 @@
+import React from "react"
+import AboutMain from "./components/AboutMain"
+import Navbar from "./components/Navbar"
+
+
+export default function AboutApi() {
+    return (
+        <div>
+            <Navbar />
+            <AboutMain />
+        </div>
+    )
+}
