+++ src/App.js
@@ -1,11 +1,16 @@
 import React, {useState, useEffect} from 'react'
-import Navbar from "./components/Navbar";
-import Dinos from "./components/Dinos";
 import { fetchDinos } from './helpers/fetchDinos';
+import Main from './Main';
+import AboutApi from './AboutApi';
+import {
+  BrowserRouter as Router, 
+  Routes, 
+  Route
+} from 'react-router-dom';
 // import Botonesnav from "./components/Botonesnav";
 
 function App() {
-  const [dinos, setdinos] = useState([])
+  //const [dinos, setdinos] = useState([])
   // const [info, setinfo] = useState({})
 
 // const onPrevious = () => {
@@ -16,7 +21,7 @@ function App() {
 //   fetchDinos(info.next)
 // }
 
-const getDinos = async () => {
+/* const getDinos = async () => {
   const baseURL = 'https://dinoapi.onrender.com/dinos';
   const dinosaur = await fetchDinos(baseURL);
   setdinos(dinosaur);
@@ -24,18 +29,15 @@ const getDinos = async () => {
 
 useEffect(() => {
   getDinos()
-}, [])
+}, []) */
 
   return (
-    <>
-        <Navbar brand="DinoAPP"/>
-
-        <div className="container mt-5">
-          {/* <Botonesnav prev={info.prev} next={info.next} onPrevious={onPrevious} onNext={onNext}/> */}
-          <Dinos dinos={dinos}/>
-          {/* <Botonesnav prev={info.prev} next={info.next} onPrevious={onPrevious} onNext={onNext}/> */}
-        </div>
-    </>
+    <Router>  
+      <Routes>
+        <Route exact path="/" element={<Main/>}/>
+        <Route exact path="aboutapi" element={<AboutApi/>}/>
+      </Routes>
+    </Router>
   );
 }
