+++ src/components/AboutMain.js
@@ -1,0 +1,23 @@
+import React from "react"
+
+export default function AboutMain() {
+    return (
+        <main>
+            <h1>
+                About
+            </h1>
+            <h2>
+                Sobre la DinoAPI
+            </h2>
+            <p>
+                Usando esta API se puede obtener información de diferentes especies de dinosaurios: su nombre, la foto del dinosaurio, su hábitat y el periodo en el que vivió.
+            </p>
+            <h1>
+                Como se obtuvo la información
+            </h1>
+            <p>
+                La información encontrada en esta API fue obtenida de INGRESAR.
+            </p>
+        </main>
+    )
+}
