+++ src/components/Navbar.js
@@ -5,6 +5,8 @@ const Navbar = ({ brand }) => {
     <nav className="navbar navbar-dark bg-dark ">
     <div className="container justify-content-center">
       <a className="navbar-brand" href="/"><img style={{width: "30px"}} alt="dino" src="favicon.png"/>{brand}</a>
+      <a href="/aboutapi" class="btn btn-info">Sobre la API</a>
+      <a href="/aboutapi" class="btn btn-info">Los creadores</a>
     </div>
   </nav>
   )
